Fix WaitForStatusRunning method
